+++ src/WithEventEmitter.test.ts
@@ -1,16 +1,16 @@
-import { WithEventPublisher } from "./WithEventPublisher";
+import { WithEventEmitter } from "./WithEventEmitter";
 
 // Sample Events interface for testing
 interface Events {
-    onFoo(a: number, b: boolean): boolean;
+    foo(a: number, b: boolean): boolean;
 }
 
-// Sample class that extends WithEventPublisher for testing
-class Widget extends WithEventPublisher<Events> {
+// Sample class that extends WithEventEmitter for testing
+class Widget extends WithEventEmitter<Events> {
     public constructor() {
-        // Helps validate that EventPublisher constructor params are passed through
+        // Helps validate that EventEmitter constructor params are passed through
         super({
-            onFoo: {
+            foo: {
                 onSubscribe: (handler) => {
                     handler(42, true);
                 },
@@ -20,36 +20,36 @@ class Widget extends WithEventPublisher<Events> {
 
     // Helps validate that the class inherits access topublishing events
     public triggerFoo(a: number, b: boolean): void {
-        this.publish.onFoo(a, b);
+        this.emit.foo(a, b);
     }
 }
 
-// Exhaustive testing is not necessary because WithEventPublisher is a simple
-// pass-through to an EventPublisher instance. We only need to cover the basics:
-// - Can subscribe via methiod inherited from WithEventPublisher.
+// Exhaustive testing is not necessary because WithEventEmitter is a simple
+// pass-through to an EventEmitter instance. We only need to cover the basics:
+// - Can subscribe via methiod inherited from WithEventEmitter.
 // - Subscription options are applied.
-// - Can unsubscribe via method inherited from WithEventPublisher.
-// - Can publish events via `publish` property inherited from WithEventPublisher.
+// - Can cancel subscriptions.
+// - Can emit events via `emit` property inherited from WithEventEmitter.
 test("General sanity check", () => {
     const widget = new Widget();
 
     const fooAlways = jest.fn();
-    const fooUnsubscribed = jest.fn();
+    const fooCancelled = jest.fn();
     const fooOnce = jest.fn();
 
-    widget.subscribe({
-        onFoo: fooAlways,
+    widget.on({
+        foo: fooAlways,
     });
 
-    // Will be unsubscribed.
+    // Will be cancelled.
     // Also testing short-hand subscription to a single event.
-    const cancelSubscription = widget.subscribe("onFoo", fooUnsubscribed);
+    const cancelSubscription = widget.on("foo", fooCancelled);
 
     // One-time only subscription
     // (validating that subscription options are passed through)
-    widget.subscribe(
+    widget.on(
         {
-            onFoo: fooOnce,
+            foo: fooOnce,
         },
         {
             once: true,
@@ -59,8 +59,8 @@ test("General sanity check", () => {
     // All handlers called initially by onSubscribe
     expect(fooAlways).toHaveBeenCalledTimes(1);
     expect(fooAlways).toHaveBeenLastCalledWith(42, true);
-    expect(fooUnsubscribed).toHaveBeenCalledTimes(1);
-    expect(fooUnsubscribed).toHaveBeenLastCalledWith(42, true);
+    expect(fooCancelled).toHaveBeenCalledTimes(1);
+    expect(fooCancelled).toHaveBeenLastCalledWith(42, true);
     expect(fooOnce).toHaveBeenCalledTimes(1);
     expect(fooOnce).toHaveBeenLastCalledWith(42, true);
 
@@ -70,8 +70,8 @@ test("General sanity check", () => {
     // Still-subscribed handlers called
     expect(fooAlways).toHaveBeenCalledTimes(2);
     expect(fooAlways).toHaveBeenLastCalledWith(1337, false);
-    expect(fooUnsubscribed).toHaveBeenCalledTimes(2);
-    expect(fooUnsubscribed).toHaveBeenLastCalledWith(1337, false);
+    expect(fooCancelled).toHaveBeenCalledTimes(2);
+    expect(fooCancelled).toHaveBeenLastCalledWith(1337, false);
     // One-time subscription not called again
     expect(fooOnce).toHaveBeenCalledTimes(1);
 
@@ -81,8 +81,8 @@ test("General sanity check", () => {
     // Still-subscribed handler called
     expect(fooAlways).toHaveBeenCalledTimes(3);
     expect(fooAlways).toHaveBeenLastCalledWith(-1, true);
-    // Unsubscribed handler not called again
-    expect(fooUnsubscribed).toHaveBeenCalledTimes(2);
+    // Cancelled handler not called again
+    expect(fooCancelled).toHaveBeenCalledTimes(2);
     // One-time subscription not called again
     expect(fooOnce).toHaveBeenCalledTimes(1);
 });