+++ src/WithEventEmitter.ts
@@ -1,4 +1,4 @@
-import { EventPublisher } from "./EventPublisher";
+import { EventEmitter } from "./EventEmitter";
 import {
     EventsOptions,
     EventsConstraint,
@@ -12,9 +12,9 @@ import {
  * Convenient base class to be inherited to make its parent a fully functional
  * {@link EventSource} implementation.
  * When inheriting from this class, your class will gain:
- * - A protected [publish]{@link WithEventPublisher#publish} property for publishing
- *   events to subscribers (just like {@link EventPublisher#publish}).
- * - A public [subscribe]{@link WithEventPublisher#subscribe} method for subscribing
+ * - A protected [emit]{@link WithEventEmitter#emit} property for emitting
+ *   events to subscribers (just like {@link EventEmitter#emit}).
+ * - A public [subscribe]{@link WithEventEmitter#subscribe} method for subscribing
  *   to your events.
  * NOTE: You must explicitly provide an interface for your events as the Events
  *       template parameter.
@@ -23,11 +23,11 @@ import {
  *           for handlers of the event.
  * @example
  * ```
- * class MyEventSource extends WithEventPublisher<{
+ * class MyEventSource extends WithEventEmitter<{
  *     // Event definitions here.
  *     // Add TSDoc comment here to explain when this event is called,
  *     // document params, etc.
- *     onNameChange(name: string): void;
+ *     nameChanged(name: string): void;
  * }> {
  *     public constructor(private name: string) {
  *         super();
@@ -35,8 +35,8 @@ import {
  *
  *     public setName(name: string): void {
  *         this.name = name;
- *         // publish the "onNameChange" event with the new name
- *         this.publish.onNameChange(name);
+ *         // emit the "nameChanged" event with the new name
+ *         this.emit.nameChanged(name);
  *     }
  * }
  *
@@ -44,46 +44,46 @@ import {
  * const myEventSource = new MyEventSource("Bob");
  *
  * // Subscribe to an event
- * myEventSource.subscribe("onNameChange", (name) => {
+ * myEventSource.on("nameChanged", (name) => {
  *     console.log(name);
  * });
  *
- * // Trigger the event to be published, which will execute the above event
+ * // Trigger the event to be emitted, which will execute the above event
  * // handler
  * myEventSource.setName("Joe");
  * ```
  */
-export class WithEventPublisher<Events extends EventsConstraint<Events>>
+export class WithEventEmitter<Events extends EventsConstraint<Events>>
     implements EventSource<Events> {
     /**
-     * The underlying implementation that handles all of the event publishing
+     * The underlying implementation that handles all of the event emitting
      * and subscribing.
      */
-    private eventPublisher: EventPublisher<Events>;
+    private eventEmitter: EventEmitter<Events>;
 
     /**
-     * A convenient proxy for publishing to all subscribed handlers of any event.
+     * A convenient proxy for emitting to all subscribed handlers of any event.
      * For each event defined by the Events interface, a method of the same name,
      * and same parameters signature, exists on this object that will call
      * all subscribed handlers of that event.
      * NOTE: The return type of every method on this proxy is `void`, regardless
      *       of the return type for the corresponding event.
      */
-    protected publish: EventPublisher<Events>["publish"];
+    protected emit: EventEmitter<Events>["emit"];
 
     /**
      * @param options - Configuration options for the events.
      */
     constructor(options?: EventsOptions<Events>) {
-        this.eventPublisher = new EventPublisher<Events>(options);
-        this.publish = this.eventPublisher.publish;
+        this.eventEmitter = new EventEmitter<Events>(options);
+        this.emit = this.eventEmitter.emit;
     }
 
     /**
      * @override
      * @inheritdoc
      */
-    public subscribe<EventName extends EventNames<Events>>(
+    public on<EventName extends EventNames<Events>>(
         eventName: EventName,
         handler: Events[EventName],
         options?: SubscriptionOptions
@@ -92,14 +92,11 @@ export class WithEventPublisher<Events extends EventsConstraint<Events>>
      * @override
      * @inheritdoc
      */
-    public subscribe(
+    public on(
         handlers: Partial<Events>,
         options?: SubscriptionOptions
     ): SubscriptionCanceller;
-    public subscribe(): SubscriptionCanceller {
-        return this.eventPublisher.subscribe.apply(
-            this.eventPublisher,
-            arguments as any
-        );
+    public on(): SubscriptionCanceller {
+        return this.eventEmitter.on.apply(this.eventEmitter, arguments as any);
     }
 }