+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+-   Renamed `EventSource.subscribe` method to `EventSource.on` (cascades down to all implementations of the interface too).
+-   Renamed `EventPublisher` to `EventEmitter`.
+    -   Renamed property `publish` to `emit`.
+-   Renamed `WithEventPublisher` to `WithEventEmitter`.
+    -   Renamed property `publish` to `emit`.
+
 ## [0.0.3] 2021-03-04
 
 ### Changed
@@ -28,6 +38,7 @@ TODO:
 -   Write good README content (sorry; source code documentation will have to suffice for now).
 -   Use rollup to bundle code?
 
-[0.0.2]: https://github.com/UselessPickles/ts-event-emitter/compare/v0.0.2...v0.0.3
+[unreleased]: https://github.com/UselessPickles/ts-event-emitter/compare/v0.0.3...HEAD
+[0.0.3]: https://github.com/UselessPickles/ts-event-emitter/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/UselessPickles/ts-event-emitter/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/UselessPickles/ts-event-emitter/tree/v0.0.1